Refactor store/conf to be at the app/theme level
