+++ packages/app/src/cli/services/environment.test.ts
@@ -18,6 +18,7 @@ import {
   DeployEnvironmentOptions,
 } from './environment.js'
 import {createExtension} from './dev/create-extension.js'
+import {CachedAppInfo, clearAppInfo, getAppInfo, setAppInfo} from './conf.js'
 import {OrganizationApp, OrganizationStore} from '../models/organization.js'
 import {updateAppIdentifiers, getAppIdentifiers} from '../models/app/identifiers.js'
 import {UIExtension} from '../models/app/extensions.js'
@@ -27,13 +28,13 @@ import metadata from '../metadata.js'
 import {loadAppName} from '../models/app/loader.js'
 import {App} from '../models/app/app.js'
 import {AllOrganizationsQuerySchemaOrganization} from '../api/graphql/all_orgs.js'
-import {store} from '@shopify/cli-kit'
 import {beforeEach, describe, expect, it, test, vi} from 'vitest'
 import {ok} from '@shopify/cli-kit/node/result.js'
 import {ensureAuthenticatedPartners} from '@shopify/cli-kit/node/session'
 import {mockAndCaptureOutput} from '@shopify/cli-kit/node/testing/output'
 
 beforeEach(() => {
+  vi.mock('./conf.js')
   vi.mock('./dev/fetch')
   vi.mock('./dev/create-extension')
   vi.mock('./dev/select-app')
@@ -45,17 +46,6 @@ beforeEach(() => {
   vi.mock('../models/app/loader.js')
   vi.mock('@shopify/cli-kit/node/session')
   vi.mocked(ensureAuthenticatedPartners).mockResolvedValue('token')
-  vi.mock('@shopify/cli-kit', async () => {
-    const cliKit: any = await vi.importActual('@shopify/cli-kit')
-    return {
-      ...cliKit,
-      store: {
-        getAppInfo: vi.fn(),
-        setAppInfo: vi.fn(),
-        clearAppInfo: vi.fn(),
-      },
-    }
-  })
 })
 
 const APP1: OrganizationApp = {
@@ -88,7 +78,7 @@ const ORG2: AllOrganizationsQuerySchemaOrganization = {
   website: '',
 }
 
-const CACHED1: store.CachedAppInfo = {appId: 'key1', orgId: '1', storeFqdn: 'domain1', directory: '/cached'}
+const CACHED1: CachedAppInfo = {appId: 'key1', orgId: '1', storeFqdn: 'domain1', directory: '/cached'}
 const STORE1: OrganizationStore = {
   shopId: '1',
   link: 'link1',
@@ -193,7 +183,7 @@ describe('ensureGenerateEnvironment', () => {
     // Given
     const input = {directory: '/app', reset: false, token: 'token'}
     vi.mocked(fetchAppFromApiKey).mockResolvedValueOnce(APP2)
-    vi.mocked(store.getAppInfo).mockReturnValue(CACHED1)
+    vi.mocked(getAppInfo).mockReturnValue(CACHED1)
 
     // When
     const got = await ensureGenerateEnvironment(input)
@@ -206,7 +196,7 @@ describe('ensureGenerateEnvironment', () => {
     const input = {directory: '/app', reset: true, token: 'token'}
     vi.mocked(fetchAppFromApiKey).mockResolvedValueOnce(APP2)
     vi.mocked(loadAppName).mockResolvedValueOnce('my-app')
-    vi.mocked(store.getAppInfo).mockReturnValue(undefined)
+    vi.mocked(getAppInfo).mockReturnValue(undefined)
 
     // When
     const got = await ensureGenerateEnvironment(input)
@@ -214,7 +204,7 @@ describe('ensureGenerateEnvironment', () => {
     // Then
     expect(got).toEqual(APP1.apiKey)
     expect(selectOrCreateApp).toHaveBeenCalledWith('my-app', [APP1, APP2], ORG1, 'token')
-    expect(store.setAppInfo).toHaveBeenCalledWith({
+    expect(setAppInfo).toHaveBeenCalledWith({
       appId: APP1.apiKey,
       title: APP1.title,
       directory: '/app',
@@ -226,7 +216,7 @@ describe('ensureGenerateEnvironment', () => {
 describe('ensureDevEnvironment', () => {
   it('returns selected data and updates internal state, without cached state', async () => {
     // Given
-    vi.mocked(store.getAppInfo).mockReturnValue(undefined)
+    vi.mocked(getAppInfo).mockReturnValue(undefined)
 
     // When
     const got = await ensureDevEnvironment(INPUT, 'token')
@@ -238,13 +228,13 @@ describe('ensureDevEnvironment', () => {
       tunnelPlugin: undefined,
       updateURLs: undefined,
     })
-    expect(store.setAppInfo).toHaveBeenNthCalledWith(1, {
+    expect(setAppInfo).toHaveBeenNthCalledWith(1, {
       appId: APP1.apiKey,
       title: APP1.title,
       directory: INPUT.directory,
       orgId: ORG1.id,
     })
-    expect(store.setAppInfo).toHaveBeenNthCalledWith(2, {
+    expect(setAppInfo).toHaveBeenNthCalledWith(2, {
       appId: APP1.apiKey,
       directory: INPUT.directory,
       storeFqdn: STORE1.shopDomain,
@@ -259,7 +249,7 @@ describe('ensureDevEnvironment', () => {
   it('returns selected data and updates internal state, with cached state', async () => {
     // Given
     const outputMock = mockAndCaptureOutput()
-    vi.mocked(store.getAppInfo).mockReturnValue(CACHED1)
+    vi.mocked(getAppInfo).mockReturnValue(CACHED1)
     vi.mocked(fetchAppFromApiKey).mockResolvedValueOnce(APP1)
     vi.mocked(fetchStoreByDomain).mockResolvedValue({organization: ORG1, store: STORE1})
 
@@ -275,13 +265,13 @@ describe('ensureDevEnvironment', () => {
     })
     expect(fetchOrganizations).not.toBeCalled()
     expect(selectOrganizationPrompt).not.toBeCalled()
-    expect(store.setAppInfo).toHaveBeenNthCalledWith(1, {
+    expect(setAppInfo).toHaveBeenNthCalledWith(1, {
       appId: APP1.apiKey,
       title: APP1.title,
       directory: INPUT.directory,
       orgId: ORG1.id,
     })
-    expect(store.setAppInfo).toHaveBeenNthCalledWith(2, {
+    expect(setAppInfo).toHaveBeenNthCalledWith(2, {
       appId: APP1.apiKey,
       directory: INPUT.directory,
       storeFqdn: STORE1.shopDomain,
@@ -292,7 +282,7 @@ describe('ensureDevEnvironment', () => {
 
   it('returns selected data and updates internal state, with inputs from flags', async () => {
     // Given
-    vi.mocked(store.getAppInfo).mockReturnValue(undefined)
+    vi.mocked(getAppInfo).mockReturnValue(undefined)
     vi.mocked(convertToTestStoreIfNeeded).mockResolvedValueOnce()
     vi.mocked(fetchAppFromApiKey).mockResolvedValueOnce(APP2)
     vi.mocked(fetchStoreByDomain).mockResolvedValue({organization: ORG1, store: STORE1})
@@ -307,7 +297,7 @@ describe('ensureDevEnvironment', () => {
       tunnelPlugin: undefined,
       updateURLs: undefined,
     })
-    expect(store.setAppInfo).toHaveBeenNthCalledWith(1, {
+    expect(setAppInfo).toHaveBeenNthCalledWith(1, {
       appId: APP2.apiKey,
       directory: INPUT_WITH_DATA.directory,
       storeFqdn: STORE1.shopDomain,
@@ -322,7 +312,7 @@ describe('ensureDevEnvironment', () => {
 
   it('throws if the store input is not valid', async () => {
     // Given
-    vi.mocked(store.getAppInfo).mockReturnValue(undefined)
+    vi.mocked(getAppInfo).mockReturnValue(undefined)
     vi.mocked(fetchAppFromApiKey).mockResolvedValueOnce(APP2)
     vi.mocked(fetchStoreByDomain).mockResolvedValue({organization: ORG1, store: undefined})
 
@@ -338,7 +328,7 @@ describe('ensureDevEnvironment', () => {
     await ensureDevEnvironment({...INPUT, reset: true}, 'token')
 
     // Then
-    expect(store.clearAppInfo).toHaveBeenCalledWith(BAD_INPUT_WITH_DATA.directory)
+    expect(clearAppInfo).toHaveBeenCalledWith(BAD_INPUT_WITH_DATA.directory)
     expect(fetchOrgAndApps).toBeCalled()
   })
 })
@@ -378,7 +368,7 @@ describe('ensureDeployEnvironment', () => {
       extensionIds: {},
     }
     vi.mocked(getAppIdentifiers).mockResolvedValue({app: undefined})
-    vi.mocked(store.getAppInfo).mockReturnValueOnce(CACHED1)
+    vi.mocked(getAppInfo).mockReturnValueOnce(CACHED1)
     vi.mocked(fetchAppFromApiKey).mockResolvedValueOnce(APP2)
     vi.mocked(ensureDeploymentIdsPresence).mockResolvedValue(identifiers)
     vi.mocked(reuseDevConfigPrompt).mockResolvedValueOnce(true)
