+++ packages/app/src/cli/services/info.test.ts
@@ -1,32 +1,23 @@
 import {info} from './info.js'
 import {fetchOrgAndApps, fetchOrganizations} from './dev/fetch.js'
 import {selectApp} from './app/select-app.js'
+import {getAppInfo} from './conf.js'
 import {AppInterface} from '../models/app/app.js'
 import {selectOrganizationPrompt} from '../prompts/dev.js'
 import {testApp, testUIExtension} from '../models/app/app.test-data.js'
 import {AppErrors} from '../models/app/loader.js'
-import {output, store} from '@shopify/cli-kit'
+import {output} from '@shopify/cli-kit'
 import {describe, it, expect, vi, beforeEach} from 'vitest'
 import {checkForNewVersion} from '@shopify/cli-kit/node/node-package-manager'
 import {ensureAuthenticatedPartners} from '@shopify/cli-kit/node/session'
 import {joinPath} from '@shopify/cli-kit/node/path'
 
 beforeEach(async () => {
+  vi.mock('./conf.js')
   vi.mock('./dev/fetch.js')
   vi.mock('./app/select-app.js')
   vi.mock('../prompts/dev.js')
   vi.mock('@shopify/cli-kit/node/session')
-  vi.mock('@shopify/cli-kit', async () => {
-    const cliKit: any = await vi.importActual('@shopify/cli-kit')
-    return {
-      ...cliKit,
-      store: {
-        getAppInfo: vi.fn(),
-        setAppInfo: vi.fn(),
-        clearAppInfo: vi.fn(),
-      },
-    }
-  })
   vi.mock('@shopify/cli-kit/node/node-package-manager')
 })
 
@@ -54,7 +45,7 @@ describe('info', () => {
       storeFqdn: 'my-app.example.com',
       updateURLs: true,
     }
-    vi.mocked(store.getAppInfo).mockReturnValue(cachedAppInfo)
+    vi.mocked(getAppInfo).mockReturnValue(cachedAppInfo)
     const app = mockApp()
 
     // When
