+++ packages/cli-kit/src/index.ts
@@ -1,3 +1,2 @@
 export * as output from './output.js'
-export * as store from './store.js'
 export * as ui from './ui.js'
