+++ packages/cli-kit/src/private/node/session/store.ts
@@ -3,7 +3,7 @@ import {keychainConstants} from '../constants.js'
 import {platformAndArch} from '../../../public/node/os.js'
 import {secureStoreSave, secureStoreFetch, secureStoreRemove} from '../secure-store.js'
 import {content, debug} from '../../../output.js'
-import {getSession, removeSession, setSession, clearAllAppInfo} from '../../../store.js'
+import {getSession, removeSession, setSession} from '../../../public/node/store.js'
 import type {Session} from './schema.js'
 
 /**
@@ -63,8 +63,6 @@ export async function remove() {
   } else {
     removeSession()
   }
-
-  clearAllAppInfo()
 }
 
 /**
